requirements.txt
Click==7.0
Flask==1.1.1
fre==0.0.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pdf2image==1.9.0
Pillow==6.1.0
pytesseract==0.3.0
Wand==0.5.7
Werkzeug==0.16.0
